feat: create uppy.tsx file
- Added a new file `src/api/uppy.tsx` to handle file uploads using Uppy.
